% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixID.R
\name{fixID}
\alias{fixID}
\title{Title}
\usage{
fixID(id)
}
\arguments{
\item{id}{a character or numeric value}
}
\value{
a character value with the appropriate number of leading zeros
}
\description{
Title
}
\examples{
ids <- c(NA, 1, 12, 234, 4567, 7899, 87357, 376256, 8765676)
fixID(ids)
}
