% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAge.R
\name{getAge}
\alias{getAge}
\title{Title}
\usage{
getAge(birthDate, refDate = Sys.Date(), unit = "year")
}
\arguments{
\item{birthDate}{A character string in the format \%Y-\%m-\%d or a date object}

\item{refDate}{A character string in the format \%Y-\%m-\%d or a date object}

\item{unit}{A character string indicating the desired unit. Must be one of 'year', 'month', 'week', or 'day'}
}
\value{
A number indicating age in the unit set in the unit argument
}
\description{
Title
}
\examples{
birthDate <- "2000-01-01"
refDate <- Sys.Date()
age <- getAge(birthDate, refDate, unit="year")
}
