% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termFunctions.R
\name{getTermFromDate}
\alias{getTermFromDate}
\title{Get the term from a date}
\usage{
getTermFromDate(date)
}
\arguments{
\item{date}{a vector with class date or datetime.}
}
\value{
a character vector of term labels.
}
\description{
A simple function to get the term (i.e. "2023F") from a date or datetime.
Relies on lubridate::semester to place a date in either the fall or spring term.
}
\examples{
date <- as.Date("2023-01-01")
getTermFromDate(date)
}
