% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termFunctions.R
\name{getTermNum}
\alias{getTermNum}
\title{Get the number of of Hampshire College term label}
\usage{
getTermNum(x)
}
\arguments{
\item{x}{a character vector of term labels in the format "2023F" where 2023 is the
year and F is the season.}
}
\value{
a numeric vector.
}
\description{
Useful for sequencing terms over time.
}
\examples{
terms <- sort(paste0(rep(2010:2020,2), c("S","F")))
getTermNum(terms)
}
