% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqyFunctions.R
\name{parseIqyUrl}
\alias{parseIqyUrl}
\title{Title}
\usage{
parseIqyUrl(path)
}
\arguments{
\item{path}{A path to a .iqy file created in Informer.}
}
\value{
String of the constructed url drawn from the .iqy file.
}
\description{
Title
}
\examples{
path <- file("https://raw.githubusercontent.com/cawenz/HCephemera/main/buildings.iqy")
url <- parseIqyUrl(path)
x <- getIqyData(url)
close(path)
}
