% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_sum.R
\name{round_sum}
\alias{round_sum}
\title{Rounding while preserving sum of vector.}
\usage{
round_sum(x, digits = 0)
}
\arguments{
\item{x}{Numeric vector; often a grouped dataframe.}

\item{digits}{Integer indicating the number of decimal places to be used while rounding.}
}
\value{
Numeric vector with rounded values that preserves the sum total.
}
\description{
Rounds a vector of numbers while preserving the sum of the vector.
Particularly useful when rounding percentages that sum to 100\%.
}
\examples{
x <- c(.33389,0.313,0.14593, 0.20217 )
round_sum(x, digits=1)
}
