% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/termFunctions.R
\name{termFromNum}
\alias{termFromNum}
\title{Get a term label from integer.}
\usage{
termFromNum(x)
}
\arguments{
\item{x}{a positive integer representing a Hampshire College term.}
}
\value{
A character string of the term (in the format 2023F
where 2023 is the year and F is the season).
}
\description{
Takes a positive integer and returns the corresponding term label.
}
\examples{
x <- 106
termFromNum(x)
}
