% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqyFunctions.R
\name{getIqyData}
\alias{getIqyData}
\title{Title}
\usage{
getIqyData(url)
}
\arguments{
\item{url}{A valid url, often the result of running parseIqyUrlthat will fetch current data from Informer.}
}
\value{
A data.frame containing the requested data. Missing values will be converted to NA.
}
\description{
Title
}
\examples{
path <- file("https://raw.githubusercontent.com/cawenz/HCephemera/main/buildings.iqy")
url <- parseIqyUrl(path)
x <- getIqyData(url)
close(path)
}
