% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqyFunctions.R
\name{getIqyData}
\alias{getIqyData}
\title{Title}
\usage{
getIqyData(url)
}
\arguments{
\item{url}{A valid url, often the result of running parseIqyUrlthat will fetch current data from Informer.}
}
\value{
A data.frame containing the requested data. Missing values will be converted to NA.
}
\description{
Title
}
\examples{
url <- "https://informer.hampshire.edu/reports/informer/query&token=fc72421f-d9eb-46e2-84a4-5a95c3d928e7&id=1687945229&locale=en_US&encoding=UTF-8&mvSelection=COMMA"
getIqyData(url)
}
